COL-1858, an eligible collaborator is any member of the course_site
